authority_protocol

Struct AuthorityState

pub struct AuthorityState<S: ProtectedStore> { /* private fields */ }

Implementations

impl<S: ProtectedStore> AuthorityState<S>

pub fn open_boot( &mut self, validated: &ValidatedRequest<OpenBootRequest>, measurement: &VerifiedBootMeasurement, ) -> Result<u64, AuthorityFault>

Allocate a protected boot epoch after authenticating the fresh challenge.

impl<S: ProtectedStore> AuthorityState<S>

pub fn restore( store: S, verified: &VerifiedProtectedRecord, boot_challenge: [u8; 32], ) -> Self

pub fn into_store(self) -> S

impl<S: ProtectedStore> AuthorityState<S>

pub fn admit_profile_validation( &mut self, validated: &ValidatedRequest<ValidateAndStageRelayProfileRequest>, ) -> Result<AdmittedProfileValidation, AuthorityFault>

pub fn stage_profile( &mut self, verified: &RootValidatedProfile, ) -> Result<RelayIntent, AuthorityFault>

impl<S: ProtectedStore> AuthorityState<S>

pub fn authorize_committed_state( &mut self, validated: &ValidatedRequest<ReadCommittedRelayStateRequest>, ) -> Result<RelayIntent, AuthorityFault>

pub fn authorize_csr_chunk( &mut self, validated: &ValidatedRequest<ReadRelayCsrChunkRequest>, ) -> Result<CsrChunkIntent, AuthorityFault>

pub fn authorize_active_public_key( &mut self, validated: &ValidatedRequest<GetRelayActivePublicKeyRequest>, clock: &impl TrustedClock, ) -> Result<RelayIntent, AuthorityFault>

impl<S: ProtectedStore> AuthorityState<S>

pub fn request_signed_time( &mut self, validated: &ValidatedRequest<RequestSignedTimeRequest>, source: &mut impl TimeChallengeSource, ) -> Result<PendingTimeChallenge, AuthorityFault>

pub fn accept_time( &mut self, verified: &VerifiedSignedTime, clock: &impl TrustedClock, ) -> Result<(), AuthorityFault>

impl<S: ProtectedStore> AuthorityState<S>

pub fn begin_enrollment( &mut self, validated: &ValidatedRequest<BeginRelayEnrollmentRequest>, clock: &impl TrustedClock, ) -> Result<EnrollmentIntent, AuthorityFault>

pub fn consume_receipt( &mut self, validated: &ValidatedRequest<ConsumeStagedRelayProfileRequest>, ) -> Result<(), AuthorityFault>

pub fn prepare_commit( &mut self, validated: &ValidatedRequest<CommitRelayGenerationRequest>, ) -> Result<PreparedCommitIntent, AuthorityFault>

pub fn record_provider_promotion( &mut self, prepared: &PreparedCommitIntent, verified: &VerifiedProviderCasReceipt, ) -> Result<(), AuthorityFault>

pub fn finalize_commit(&mut self) -> Result<(), AuthorityFault>

pub fn abort( &mut self, validated: &ValidatedRequest<AbortRelayEnrollmentRequest>, ) -> Result<u64, AuthorityFault>

pub fn authorize_tls_signature( &mut self, validated: &ValidatedRequest<SignTls13ClientCertificateVerifyRequest>, clock: &impl TrustedClock, ) -> Result<TlsSignatureIntent, AuthorityFault>

impl<S: ProtectedStore> AuthorityState<S>

pub fn authorize_profile_upload( &mut self, validated: &ValidatedRequest<BeginRelayProfileUploadRequest>, ) -> Result<ProfileUploadIntent, AuthorityFault>

pub fn acknowledge_profile_upload( &mut self, intent: &ProfileUploadIntent, ) -> Result<ProfileUploadIntent, AuthorityFault>

pub fn authorize_profile_chunk( &mut self, validated: &ValidatedRequest<WriteRelayProfileChunkRequest>, ) -> Result<ProfileChunkIntent, AuthorityFault>

pub fn acknowledge_profile_chunk( &mut self, intent: &ProfileChunkIntent, ) -> Result<ProfileUploadIntent, AuthorityFault>

pub fn reject_profile_upload( &mut self, intent: &ProfileUploadIntent, ) -> Result<(), AuthorityFault>

pub fn reject_profile_chunk( &mut self, intent: &ProfileChunkIntent, ) -> Result<(), AuthorityFault>

impl<S: ProtectedStore> AuthorityState<S>

pub const fn mode(&self) -> AuthorityMode

Return the absorbing authority mode without exposing mutation.

pub const fn boot_state(&self) -> BootState

Return the current protected boot state.

pub const fn opened_boot_fact(&self) -> Option<OpenedBootFact>

Return the signed response fact for the currently open measured boot.

pub const fn time_state(&self) -> TimeState

Return the current one-shot time lease state.

pub const fn relay_state(&self) -> RelayProfileState

Return the current relay transaction snapshot.

pub const fn time_floors(&self) -> ProtectedTimeFloors

Return retained signed-time rollback floors for durable persistence.

impl<S: ProtectedStore> AuthorityState<S>

pub const fn new(store: S, config: AuthorityStateConfig) -> Self